Add components for the feature regarding the void
When falling into the void, the mecha
gets back on the last platform it landed on.
PlatformMemorizer must be attached to the
Mecha GameObject.
VoidMechaFollower must be given the Mecha
GameObject.
An extra tag "Platform" is necessary for
them to be memorized and therefore for the
feature to work properly.